requirements.txt
pandas>=2.0.0
yfinance>=0.2.0
akshare>=1.0.0
# google-search-results>=2.4.0
dashscope>=1.14.0
requests>=2.31.0
python-dotenv>=1.0.0
openbb
feedparser
newspaper3k
lxml[html_clean]
newsapi-python
google-search-results
sentence-transformers
